moved config logger to helper.py
